+++ deepmatch/layers/__init__.py
@@ -1,16 +1,15 @@
 from deepctr.layers import custom_objects
 from deepctr.layers.utils import reduce_sum
 
-from .core import PoolingLayer, SampledSoftmaxLayer, Similarity, LabelAwareAttention, CapsuleLayer,SampledSoftmaxLayerv2,EmbeddingIndex
+from .core import PoolingLayer, Similarity, LabelAwareAttention, CapsuleLayer,SampledSoftmaxLayer,EmbeddingIndex
 from ..utils import sampledsoftmaxloss
 
 _custom_objects = {'PoolingLayer': PoolingLayer,
-                   'SampledSoftmaxLayer': SampledSoftmaxLayer,
                    'Similarity': Similarity,
                    'LabelAwareAttention': LabelAwareAttention,
                    'CapsuleLayer': CapsuleLayer,
                    'reduce_sum':reduce_sum,
-                   'SampledSoftmaxLayerv2':SampledSoftmaxLayerv2,
+                   'SampledSoftmaxLayer':SampledSoftmaxLayer,
                    'sampledsoftmaxloss':sampledsoftmaxloss,
                    'EmbeddingIndex':EmbeddingIndex
                    }