+++ docs/source/History.md
@@ -1,3 +1,3 @@
 # History
-- 04/10/2020 : DeepMatch first version  is released on [PyPi](https://pypi.org/project/deepmatch/)
+- 04/10/2020 : [v0.1.2](https://github.com/shenweichen/DeepMatch/releases/tag/v0.1.2) released.Support [saving and loading model](./FAQ.html#save-or-load-weights-models). 
 - 04/06/2020 : DeepMatch first version  is released on [PyPi](https://pypi.org/project/deepmatch/)